requirements.txt
flask
Gunicorn
google-cloud-language>=2.0.0,<3.0.0